feat(std-json-api): refactoring resource and data adapter api
the adapter api simplified in favour of multiple queries

BREAKING CHANGE: apater api changed
