+++ packages/std-json-api/json-api.ts
@@ -865,52 +865,3 @@ export type JsonApiDocument =
   | JsonApiCollectionDocument
   | JsonApiErrorDocument
   | JsonApiMetaDocument;
-
-/// ===========================================================================
-// Resource Descriptor
-// ============================================================================
-
-/**
- * An attribute or relationship type for a JSON:API resource.
- */
-export type AttributeType =
-  | "boolean"
-  | "string"
-  | "number"
-  | "date"
-  | "object"
-  | "array"
-  | "unknown";
-
-/**
- * A mapping of attribute names to their types.
- */
-export type DescriptorAttributes = Record<string, AttributeType>;
-
-// export type Relationships = Record<string, { resource: string; idKey: string }>;
-
-export type ResourceDescriptor = {
-  /**
-   * The unique identifier for the resource type.
-   * E.g., 'articles', 'users', etc.
-   */
-  resourceId: string;
-
-  /**
-   * The base URL for the resource.
-   * E.g., 'https://api.example.com/articles'
-   */
-  baseUrl: string;
-
-  /**
-   * The primary key for the resource.
-   * E.g., 'id' or ['firstName', 'lastName'] for composite keys.
-   */
-  idKey: string | string[];
-
-  /**
-   * The attributes of the resource.
-   * E.g., { title: 'string', content: 'string', publishedAt: 'date' }
-   */
-  attributes: DescriptorAttributes;
-};
